feat: optimize code by using Lombok & Mapstruct - day3

+++ src/main/java/com/shongon/identity_service/service/UserService.java
@@ -2,56 +2,76 @@ package com.shongon.identity_service.service;
 
 import com.shongon.identity_service.dto.request.CreateUserRequest;
 import com.shongon.identity_service.dto.request.UpdateUserRequest;
+import com.shongon.identity_service.dto.response.CreateUserResponse;
+import com.shongon.identity_service.dto.response.GetAllUsersResponse;
+import com.shongon.identity_service.dto.response.UpdateUserResponse;
+import com.shongon.identity_service.dto.response.ViewUserResponse;
 import com.shongon.identity_service.entity.User;
 import com.shongon.identity_service.exception.AppException;
 import com.shongon.identity_service.exception.ErrorCode;
+import com.shongon.identity_service.mapper.UserMapper;
 import com.shongon.identity_service.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class UserService {
-    @Autowired
-    private UserRepository userRepository;
+    private final UserRepository userRepository;
+    private final UserMapper userMapper;
 
-    public User createUser(CreateUserRequest request) {
-        User user = new User();
+    @Transactional
+    public CreateUserResponse createUser(CreateUserRequest request) {
+        validateUsername(request.getUsername());
 
-        if(userRepository.existsByUsername(request.getUsername()))
-            throw new AppException(ErrorCode.USER_EXISTED);
-
-        user.setUsername(request.getUsername());
-        user.setPassword(request.getPassword());
-        user.setFirstName(request.getFirstName());
-        user.setLastName(request.getLastName());
-        user.setBirthDate(request.getBirthDate());
-
-        return userRepository.save(user);
+       return Optional.of(request)
+               .map(userMapper::createUser)
+               .map(userRepository::save)
+               .map(userMapper::toCreateUserResponse)
+               .orElseThrow(() ->  new AppException(ErrorCode.USER_CREATION_FAILED));
     }
 
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
+    @Transactional
+    public List<GetAllUsersResponse> getAllUsers() {
+        return userRepository.findAll()
+                .parallelStream()
+                .map(userMapper::toGetAllUserResponse)
+                .toList();
     }
 
-    public User getUserById(String id) {
+    @Transactional
+    public ViewUserResponse getUserById(String id) {
         return userRepository.findById(id)
+                .map(userMapper::toViewUserResponse)
                 .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
     }
 
-    public User updateUser(String userId, UpdateUserRequest request) {
-        User user = getUserById(userId);
-
-        user.setPassword(request.getPassword());
-        user.setFirstName(request.getFirstName());
-        user.setLastName(request.getLastName());
-        user.setBirthDate(request.getBirthDate());
-
-        return userRepository.save(user);
+    @Transactional
+    public UpdateUserResponse updateUser(String userId, UpdateUserRequest request) {
+        return userRepository.findById(userId)
+                .map(user ->
+                {
+                    userMapper.updateUser(user,request);
+                    return userRepository.save(user);
+                })
+                .map(userMapper::toUpdateUserResponse)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
     }
 
+    @Transactional
     public void deleteUser(String userId) {
+        if (!userRepository.existsById(userId))
+            throw new AppException(ErrorCode.USER_NOT_FOUND);
+
         userRepository.deleteById(userId);
+    }
+
+    private void validateUsername(String username) {
+        if (userRepository.existsByUsername(username))
+            throw new AppException(ErrorCode.USER_EXISTED);
     }
 }
